+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Body task targets can be read directly from a configuration
 - Example to load a custom URDF description
 - Example: Upkie crouching, visualized in yourdfpy
+- Generalize configuration limits to any root joint
 - Handle descriptions that have no velocity limit
 - Posture task targets can be read directly from a configuration
 - Simple rate limiter in ``pink.utils``